requirements.txt
discord.py==1.7.3
psycopg2==2.9.1